change to example folder
